refactor(r-checkbox.vue): updating  method name and adding comments how the component works
The component itself does not contain any state - it calculates it based on the props the parent is
sending. Clicking on it, should trigger state update, but since there is no state we try to
calculate the new one, and emit it to the parent so it updates the checkbox component. The method
was named "getModelValue", which was misleading as it was exactly the OPPOSITE of the current
model's value.